+++ parity-crypto/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog].
 [Keep a Changelog]: http://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
+- License changed from GPL3 to dual MIT/Apache2. [#342](https://github.com/paritytech/parity-common/pull/342)
 
 ## [0.5.0] - 2020-02-08
 - Remove `inv()` from `SecretKey` (breaking) (https://github.com/paritytech/parity-common/pull/258)